router and call history resetting
