--- packages/core/src/__tests__/FetchMock/route-creation.test.js
@@ -1,8 +1,0 @@
-import { describe, expect, it } from 'vitest';
-import fetchMock from '../../FetchMock';
-
-import Route from '../../Route.js';
-
-describe('Route creation', () => {
-	describe('FetchMock.route()', () => {});
-});
